backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
httpx>=0.27.0
selectolax>=0.3.17
playwright>=1.43.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
python-dateutil>=2.9.0
aiofiles>=23.2.1
